Add function to insert meshblock data into a global array
This currently works only in unigrid. It should be possible to extend
this to work for mesh refinement by taking into account the level of
each block and returning a trivial slicing for blocks at at a level
other than that being filled.

This also makes the way coordinate arrays are returned consistent with
the underlying indexing of the data.